+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Piwik Platform Changelog
 
-This is a changelog for Piwik platform developers. All changes for our HTTP API's, Plugins, Themes, etc will be listed here. 
+This is a changelog for Piwik platform developers. All changes for our HTTP API's, Plugins, Themes, etc will be listed here.
+
+## Piwik 2.11.0
+
+### Deprecations
+* The following methods of `Piwik\Common` have been deprecated in favor of the new `Piwik\Intl` component:
+  * `getContinentsList()`: use `RegionDataProvider::getContinentList()`
+  * `getCountriesList()`: use `RegionDataProvider::getCountryList()`
+  * `getLanguagesList()`: use `LanguageDataProvider::getLanguageList()`
+  * `getLanguageToCountryList()`: use `LanguageDataProvider::getLanguageToCountryList()`
 
 ## Piwik 2.11.0